Create a protocol for PrintFactory / ConsolePrintFactory
